requirements copy.txt
fastapi
uvicorn
langchain-google-genai
langgraph
langchain
googlemaps
google-cloud-bigquery
requests
parsedatetime
pytz
python-dotenv
langserve[all]